requirements.txt
asgiref==3.8.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
Markdown==3.7
mysqlclient==2.2.6
py3dns==4.0.2
sqlparse==0.5.3
tzdata==2024.2
